+++ android/src/main/java/com/opentokreactnative/OTSessionManager.java
@@ -338,7 +338,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
             Stream mStream = streams.get(streamId);
             if (mStream == null) {
                 promise.reject("Stream not found.");
-                return;
+                continue;
             }
             mExcludedStreams.add(mStream);
         }
@@ -865,14 +865,14 @@ public class OTSessionManager extends ReactContextBaseJavaModule
         printLogs("onStreamDropped: Stream Dropped: "+stream.getStreamId() +" in session: "+session.getSessionId());
     }
     @Override
-    public void onMuteForced​(Session session, MuteForcedInfo info) {
+    public void onMuteForced(Session session, MuteForcedInfo info) {
 
         WritableMap muteForcedInfo = Arguments.createMap();
         String sessionId = session.getSessionId();
         muteForcedInfo.putString("sessionId", sessionId);
         Boolean active = info.getActive();
         muteForcedInfo.putBoolean("active", active);
-        sendEventMap(this.getReactApplicationContext(), session.getSessionId() + ":" + sessionPreface + "onMuteForced​", muteForcedInfo);
+        sendEventMap(this.getReactApplicationContext(), session.getSessionId() + ":" + sessionPreface + "onMuteForced", muteForcedInfo);
         printLogs("Mute forced -- active: " + active + " in session: " + sessionId);
     }
 
@@ -968,7 +968,7 @@ public class OTSessionManager extends ReactContextBaseJavaModule
     }
 
     @Override
-    public void onMuteForced​(PublisherKit publisher) {
+    public void onMuteForced(PublisherKit publisher) {
 
         String publisherId = Utils.getPublisherId(publisher);
         if (publisherId.length() > 0) {