+++ ios/OpenTokReactNative/OTSessionManager.swift
@@ -417,12 +417,25 @@ class OTSessionManager: RCTEventEmitter {
         resolve(supportedCodecs)
     }
     
-    @objc func forceMuteAll(_ sessionId: String, excludedStreamIds: NSArray, resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) -> Void{
+    @objc func forceMuteAll(_ sessionId: String, excludedStreamIds: Array<String>, resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) -> Void{
         guard let session = OTRN.sharedState.sessions[sessionId] else {
             reject("event_failure", "Session ID not found", nil)
             return
         }
-        return resolve(true);
+        var excludedStreams:[OTStream] = []
+        for streamId in excludedStreamIds {
+            guard let stream = OTRN.sharedState.subscriberStreams[streamId] ?? OTRN.sharedState.publisherStreams[streamId] else {
+                continue // Ignore bogus stream IDs
+            }
+            excludedStreams.append(stream)
+        }
+        var error: OTError?
+        session.forceMuteAll(excludedStreams, error: &error)
+        if let error = error {
+          reject("event_failure", error.localizedDescription, nil)
+          return
+        }
+        return resolve(true)
     }
 
     @objc func forceMuteStream(_ sessionId: String, streamId: String, resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) -> Void{
@@ -619,7 +632,7 @@ extension OTSessionManager: OTSessionDelegate {
     func session(_ session: OTSession, info muteForced: OTMuteForcedInfo) {
         var muteForcedInfo: Dictionary<String, Any> = [:];
         muteForcedInfo["active"] = muteForced.active;
-        self.emitEvent("\(session.sessionId):\(EventUtils.sessionPreface)muteFoced", data: muteForcedInfo)
+        self.emitEvent("\(session.sessionId):\(EventUtils.sessionPreface)muteForced", data: muteForcedInfo)
         printLogs("OTRN Session: Session muteForced - active:  \(muteForced.active)")
     }
 }